Refactor - Entities from synthetic.py are put in separate files: . camera.py . jacobian_tools.py . sampling.py . scene.py . scene_objects.py - File synthetic.py contains synthetic pipeline for testing purposes - Changed testing code accordingly
